+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## v1.4.0
+
+_2016-06-06_
+
+- Prompt views being rendered by Android Studio's Layout Editor Preview no longer cause an `IllegalStateException` (issue #152);
+- Users can now customize more attributes of the `DefaultLayoutPromptView` in both xml and code:
+  - Text size;
+  - Button border width;
+  - Button corner radius.
+
 ## v1.3.1
 
 _2016-04-12_